Module deathframe_physics::collision::collision_check

Module containing collision checking functions.

Functions

do_rect_ids_match
Checks if the given IDs are the same.

do_rect_tags_match
Checks if the first given tag should collide with the second given tag.

do_rects_collide
Returns true if the two passed CollisionRects are in collision; also checks, that their entity IDs are not the same, and that their tags allow them to collide with each other.

do_rects_intersect
Returns true if the two passed Rects intersect with each other.