Trait deathframe_physics::collision::tag::CollisionTag

pub trait CollisionTag: Send + Sync + Clone + PartialEq {
    fn collides_with(&self, other: &Self) -> bool;
}

This tag is used for collision checking and when moving with solids. Implement this trait for your own type to use with Collider and Collidable.

Required methods

fn collides_with(&self, other: &Self) -> bool

Loading content...