Module deathframe_physics::systems::prelude

Structs

ApplyBaseFrictionSystem
Constantly applies friction to entities with BaseFriction, for each axis. Only if friction is enabled for the axis (see BaseFriction).

ApplyGravitySystem

MoveEntitiesSystem
This system is responsible for moving all entities with Transform and Velocity, by manipulating their Transform appropriately. It also handles collision with Solid entities; Solid entities may not move into each other.

UpdateCollisionsSystem
The UpdateCollisionsSystem is in charge of setting collision states for colliding entities. Entities with CheckCollision (and with Collision) check for collision against other entities with Collision. Only checks for entities with either NO Loadable and NO Loaded components or for entities with Loadable AND Loaded components; does not check for entities with Loadable but NOT Loaded components.